requirements.txt
# Base

annotated-types==0.7.0
anyio==4.7.0
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
pydantic==2.10.4
pydantic_core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
dlib-bin==19.24.6
python-dotenv==1.0.1
python-socketio==5.11.2
requests==2.32.3

# Interview Analytics

opencv-python==4.10.0.84
opencv-contrib-python==4.11.0.86
openpyxl==3.1.0
sounddevice==0.4.7
numpy==1.26.4
soundfile==0.12.1
pydub==0.25.1
tensorflow==2.17.0
pandas==2.1.4
librosa==0.10.2.post1
SpeechRecognition==3.10.4
resampy==0.4.3
hmmlearn==0.3.2
eyed3==0.9.7
imblearn==0.0
plotly==5.19.0
mediapipe==0.10.14
pyAudioAnalysis==0.3.14
torch==2.3.1
torchvision==0.18.1
timm==1.0.3
pillow==11.0.0
moviepy==1.0.3
transformers==4.42.4
nltk==3.9.1
wordcloud==1.9.3
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
accelerate==1.0.0
tf_keras==2.17.0
keras==3.4.1

# Voice Module
openai-whisper==20240930
silero-vad==5.1.2
vaderSentiment==3.3.2
textblob==0.19.0
seaborn==0.13.2